refactor: moving locks into the internal package
